back_end/brandi-project/requirements.txt
alembic==1.4.3
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
protobuf==3.14.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
